BUGFIX: all the mouseclicks inside the editscreen didn't work  status:  fixed BUGFIX: wallbrush did something weird.   status: fixed
wallbrush is now acctuately a brush instaid of a upgraded pencil tool

WORKING BUILD

+++ src/sinius/maze/gameEngine/GeneralListener.java
@@ -90,13 +90,13 @@ public class GeneralListener implements ComponentListener, KeyListener, MouseLis
 
 	@Override
 	public void keyPressed(KeyEvent e) {
-		if(!pressedKeys.contains(e.getKeyCode()))
-			pressedKeys.add(e.getKeyCode());
+		if(!pressedKeys.contains((Integer)e.getKeyCode()))
+			pressedKeys.add((Integer)e.getKeyCode());
 	}
 
 	@Override
 	public void keyReleased(KeyEvent e) {
-		if(pressedKeys.contains(e.getKeyCode()))
+		if(pressedKeys.contains((Integer)e.getKeyCode()))
 			pressedKeys.remove((Integer)e.getKeyCode());
 	}
